requirements.txt
streamlit==1.27.0
Pillow>=7.1.0,<10.0.0
